==> Code/2.Distance_computation.ipynb <==
import pandas as pd
import numpy as np
from haversine import Unit
from haversine import haversine as hs


### filepaths
primary_facility_path = '../Data/Primary_Facility_50.csv'
secondary_facility_path = '../Data/Secondary_Facility_500.csv'
supplier_path = '../Data/Supplier.csv'
demand_path = '../ResearchData/1.Demand_Distribution.csv'
distance_path = '../Data/DISTANCE_DICT_VP50_VQ_500.pkl'

demand_df = pd.read_csv(demand_path)
if demand_df.isna().sum().sum()!=0:
    print('ERROR!!!')

DEMAND = tuple(demand_df['Demand'])
CUSTOMER = tuple(demand_df['Index'])

DEMAND_DICT = dict(zip(CUSTOMER, DEMAND))

q_df = pd.read_csv(secondary_facility_path)
p_df = pd.read_csv(primary_facility_path)
s_df = pd.read_csv(supplier_path)


q_df['Index'] = q_df.index
s_df['Index'] = s_df.index
p_df['Index'] = p_df.index
demand_df['Index'] = demand_df.index

DISTANCE_DICT = {}
DISTANCE_DICT['SP'] = {}
DISTANCE_DICT.update({'PQ' : {}, 'PP_' : {}, 'QK' : {}})

%%time
### DISTANCE BETWEEN P and P_ and vice-versa

for i in range(len(p_df)):
    for j in range(i + 1, len(p_df)):
        dist = hs((p_df.at[i, 'Latitude'], p_df.at[i, 'Longitude']), (p_df.at[j, 'Latitude'], p_df.at[j, 'Longitude']), unit='mi')
        DISTANCE_DICT['PP_'].update({f'XPP_{i}_{j}' : dist, f'XPP_{j}_{i}' : dist})

%%time
### DISTANCE BETWEEN S and P

for i in range(len(s_df)):
    for j in range(len(p_df)):
        dist = hs((s_df.at[i, 'Latitude'], s_df.at[i, 'Longitude']), (p_df.at[j, 'Latitude'], p_df.at[j, 'Longitude']), unit='mi')
        DISTANCE_DICT['SP'].update({f'XSP_{i}_{j}' : dist})

%%time
### DISTANCE BETWEEN P and Q

for i in range(len(p_df)):
    for j in range(len(q_df)):
        dist = hs((p_df.at[i, 'Latitude'], p_df.at[i, 'Longitude']), (q_df.at[j, 'Latitude'], q_df.at[j, 'Longitude']), unit='mi')
        DISTANCE_DICT['PQ'].update({f'XPQ_{i}_{j}' : dist})

%%time
### DISTANCE BETWEEN Q and K

for i in range(len(q_df)):
    for j in range(len(demand_df)):
        dist = hs((q_df.at[i, 'Latitude'], q_df.at[i, 'Longitude']), (demand_df.at[j, 'Latitude'], demand_df.at[j, 'Longitude']), unit='mi')
        DISTANCE_DICT['QK'].update({f'XQK_{i}_{j}' : dist})

import pickle

with open(distance_path, 'wb') as fp:
    pickle.dump(DISTANCE_DICT, fp) 

DISTANCE_DICT['SP']
