==> Code/4.MILP_implementation_Pulp.ipynb <==
import pandas as pd
from pulp import *
import pickle

### filepaths
primary_facility_path = '../Data/Primary_Facility_50.csv'
secondary_facility_path = '../Data/Secondary_Facility_500.csv'
supplier_path = '../Data/Supplier.csv'
demand_path = '../ResearchData/1.Demand_Distribution.csv'
distance_path = '../Data/DISTANCE_DICT_VP50_VQ_500.pkl'

# fetch input data
p_df = pd.read_csv(primary_facility_path)
q_df = pd.read_csv(secondary_facility_path)
s_df = pd.read_csv(supplier_path)
demand_df = pd.read_csv(demand_path)

q_df['Index'] = q_df.index
s_df['Index'] = s_df.index
p_df['Index'] = p_df.index
demand_df['Index'] = demand_df.index


p_df['Index'] = p_df['Index'].apply(lambda x : 'P'+ str(x))
q_df['Index'] = q_df['Index'].apply(lambda x : 'Q'+ str(x))
s_df['Index'] = s_df['Index'].apply(lambda x : 'S'+ str(x))
demand_df['Index'] = demand_df['Index'].apply(lambda x : 'K' + str(x))

with open(distance_path, 'rb') as fp:
    DISTANCE_DICT = pickle.load(fp) 

## SETS
PRIMARY_FACILITY = tuple(p_df['Index'])
SECONDARY_FACILITY =  tuple(q_df['Index'])
CUSTOMER = tuple(demand_df['Index'])
SUPPLIER = tuple(s_df['Index'])


## VARIABLES
SUPPLIER_TO_PRIMARY = 'Xsp'
PRIMARY_TO_SECONDARY = 'Ypq'
PRIMARY_TO_PRIMARY = 'Tpp'
SECONDARY_TO_CUSTOMER = 'Zqk'
FACILITY = 'G'

## PARAMETERS
DEMAND = tuple(demand_df['Demand'])
DISTANCE = DISTANCE_DICT

FIXED_COST_PRIMARY = 250000
FIXED_COST_SECONDARY = 25000
PACKAGING_COST_PRIMARY = 0.035
PACKAGING_COST_SECONDARY = 0.140

R_PRIMARY = 2
R_SECONDARY = 50
C_SUPPLIER = 0.0006
C_PRIMARY = 0.0024
C_CUSTOMER = 0.012


DEMAND_DICT = dict(zip(CUSTOMER, DEMAND))

# # model
# model = grb.Model("MILP")

# Create a model
model = LpProblem("MILP", sense=LpMinimize)  # You can use LpMaximize for maximization problems


## Define decision variables
L = 0.5
# L = LpVariable(name="L", lowBound=0, cat=LpContinuous)
Xsp_var = LpVariable.dicts(SUPPLIER_TO_PRIMARY, [(s, p) for s in SUPPLIER for p in PRIMARY_FACILITY], lowBound=0, cat='Continuous')
Ypq_var = LpVariable.dicts(PRIMARY_TO_SECONDARY, [(p, q) for p in PRIMARY_FACILITY for q in SECONDARY_FACILITY], lowBound=0, cat='Continuous')
Zqk_var = LpVariable.dicts(SECONDARY_TO_CUSTOMER, [(q, k) for q in SECONDARY_FACILITY for k in CUSTOMER], lowBound=0, cat='Continuous')
Tpp_var = LpVariable.dicts(PRIMARY_TO_PRIMARY, [(p1, p2) for p1 in PRIMARY_FACILITY for p2 in set(PRIMARY_FACILITY)-set([p1])], lowBound=0, cat='Continuous')

G_var = {('p', p): LpVariable(name=FACILITY + '_p_' + str(p), cat='Binary') for p in PRIMARY_FACILITY}
G_var.update({('q', q): LpVariable(name=FACILITY + '_q_' + str(q), cat='Binary') for q in SECONDARY_FACILITY})


## Add constraints for R_P and R_Q
model += lpSum(G_var[('p',p)] for p in PRIMARY_FACILITY) == R_PRIMARY
model += lpSum(G_var[('q',q)] for q in SECONDARY_FACILITY) == R_SECONDARY


for p in PRIMARY_FACILITY:
    model += lpSum(Xsp_var[(s, p)] for s in SUPPLIER) <= lpSum(DEMAND_DICT[k] * G_var[('p', p)] for k in CUSTOMER)
    model += lpSum(Ypq_var[(p, q)] for q in SECONDARY_FACILITY) <= lpSum(DEMAND_DICT[k] * G_var[('p', p)] for k in CUSTOMER)

for q in SECONDARY_FACILITY:
    model += lpSum(Ypq_var[(p, q)] for p in PRIMARY_FACILITY) <= lpSum(DEMAND_DICT[k] * G_var[('q', q)] for k in CUSTOMER)
    model += lpSum(Zqk_var[(q, k)] for k in CUSTOMER) <= lpSum(DEMAND_DICT[k] * G_var[('q', q)] for k in CUSTOMER)


for p in PRIMARY_FACILITY:
    model += lpSum(Tpp_var[(p, p_)] for p_ in set(PRIMARY_FACILITY)-set([p])) <= lpSum(DEMAND_DICT[k] * G_var[('p', p)] for k in CUSTOMER)

for p_ in PRIMARY_FACILITY:
    model += lpSum(Tpp_var[(p, p_)] for p in set(PRIMARY_FACILITY)-set([p_])) <= lpSum(DEMAND_DICT[k] * G_var[('p', p_)] for k in CUSTOMER)


for p in PRIMARY_FACILITY:
    model += lpSum(Xsp_var[(s, p)] for s in SUPPLIER) + lpSum(Tpp_var[(p_, p)] for p_ in set(PRIMARY_FACILITY)-set([p])) == lpSum(Ypq_var[(p, q)] for q in SECONDARY_FACILITY) + lpSum(Tpp_var[(p, p_)] for p_ in set(PRIMARY_FACILITY)-set([p]))


for q in SECONDARY_FACILITY:
    model += lpSum(Ypq_var[(p, q)] for p in PRIMARY_FACILITY) == lpSum(Zqk_var[(q, k)] for k in CUSTOMER)


for p in PRIMARY_FACILITY:
    model += lpSum(Tpp_var[(p_, p)] for p_ in set(PRIMARY_FACILITY)-set([p])) == L * lpSum(Ypq_var[(p, q)] for q in SECONDARY_FACILITY)


for k in CUSTOMER:
    model += lpSum(Zqk_var[(q, k)] for q in SECONDARY_FACILITY) == DEMAND_DICT[k]


model.sense = LpMinimize

objective = lpSum(C_SUPPLIER * DISTANCE_DICT['SP'][f'XSP_{s[1:]}_{p[1:]}'] * Xsp_var[(s, p)] for s in SUPPLIER for p in PRIMARY_FACILITY) + \
            lpSum(C_PRIMARY * DISTANCE_DICT['PQ'][f'XPQ_{p[1:]}_{q[1:]}'] * Ypq_var[(p, q)] for p in PRIMARY_FACILITY for q in SECONDARY_FACILITY) + \
            lpSum(C_PRIMARY * DISTANCE_DICT['PP_'][f'XPP_{p[1:]}_{p_[1:]}'] * Tpp_var[(p, p_)] for p in PRIMARY_FACILITY for p_ in set(PRIMARY_FACILITY)-set([p])) + \
            lpSum(C_CUSTOMER * DISTANCE_DICT['QK'][f'XQK_{q[1:]}_{k[1:]}'] * Zqk_var[(q, k)] for q in SECONDARY_FACILITY for k in CUSTOMER) + \
            lpSum(FIXED_COST_PRIMARY * G_var[('p', p)] for p in PRIMARY_FACILITY) +\
            lpSum(FIXED_COST_SECONDARY * G_var[('q', q)] for q in SECONDARY_FACILITY) + \
            lpSum(PACKAGING_COST_PRIMARY * (lpSum(Ypq_var[(p, q)] for q in SECONDARY_FACILITY) + lpSum(Tpp_var[(p, p_)] for p_ in set(PRIMARY_FACILITY)-set([p]))) for p in PRIMARY_FACILITY) + \
            lpSum(PACKAGING_COST_SECONDARY * (lpSum(Zqk_var[(q, k)] for k in CUSTOMER)) for q in SECONDARY_FACILITY)


model.setObjective(objective)

LpProblem.writeLP(model, '../pulp_exps/MILP_pulp.lp')


# # Get the number of variables
# num_variables = len(model.getVars())

# # Get the number of constraints
# num_constraints = len(model.getConstrs())

num_variables = len(model.variables())
num_constraints = len(model.constraints)

print(f'# Var : {num_variables}')
print(f'# Constraints : {num_constraints}')

model.status

LpSolverDefault.msg = 1

# model.optimize()
model.solve(PULP_CBC_CMD(msg=1, timeLimit=1000))


for p in range(50):
    if pulp.value(Xsp_var['S0', 'P'+str(p)])!=0:
        print(p, pulp.value(Xsp_var['S0', 'P'+str(p)]))
        

for p in range(50):
    for q in range(500):
        if pulp.value(Ypq_var[f'P{p}', f'Q{q}']) != 0:
            print(p, q, pulp.value(Ypq_var[f'P{p}', f'Q{q}']))


for q in range(500):
    for k in range(371):
        if pulp.value(Zqk_var[f'Q{q}', f'K{k}']) != 0:
            print(q, k, pulp.value(Zqk_var[f'Q{q}', f'K{k}']))


for p in range(50):
    for p_ in range(50):
        if p!=p_:
            if pulp.value(Tpp_var[f'P{p}', f'P{p_}']) != 0:
                print(p, p_, pulp.value(Tpp_var[f'P{p}', f'P{p_}']))
