==> Code/5.Agglomerative_clustering.ipynb <==
import pandas as pd
import numpy as np
from haversine import haversine as hs
from haversine import Unit
import pickle

###----------------------FILEPATHS----------------------###
secondary_facility_path = '../Data/Secondary_Facility_500.csv'
demand_path = '../ResearchData/1.Demand_Distribution.csv'

# result_path = '../Results/Agglomerative_clustering_rq_200.pkl'
result_path = '../Results/Agglomerative_clustering_rq_50.pkl'

customer_path = '../Results/50_customer_coordinates.npy'
# customer_path = '../Results/200_customer_coordinates.npy'

# optimized_demand_path = '../Results/200_customer_demands.npy'
optimized_demand_path = '../Results/50_customer_demands.npy'

###----------------------PARAMETERS----------------------###
# Rq = 200
Rq = 50
stdev = 100

###----------------------READ ALL INPUT DATAFILES----------------------###
q_df = pd.read_csv(secondary_facility_path)
demand_df = pd.read_csv(demand_path)
demand_df['Co-ordinates'] = demand_df.apply(lambda x : np.array([x['Latitude'], x['Longitude']]), axis=1)

# Function to compute similarity matrix using demand, coordinates, distance type
def get_similarity_matrix(W, demand, stdev, distance_type='haversine'):
    if distance_type=='euclidean':
        ## Distance matrix using Euclidean distance
        distance_matrix = np.linalg.norm(W[:, np.newaxis, :] - W[np.newaxis, :, :], axis=-1)
        n = len(W)
        similarity_matrix = np.zeros((n, n))

        for i in range(n):
            for j in range(n):
                if i == j:
                    similarity_matrix[i, j] = 0
                else:
                    similarity_matrix[i, j] = np.mean([demand[i], demand[j]]) * np.exp(-distance_matrix[i, j]/stdev**2)

        ## Putting lower triangle to zero
        n = similarity_matrix.shape[0]
        similarity_matrix[np.tril_indices(n)] = 0

    else:
        ## Distance matrix using Haversine distance
        n = len(W)
        distance_matrix = np.zeros((n, n))
        for i in range(n):
            for j in range(n):
                distance_matrix[i, j] = hs(tuple(W[i]), tuple(W[j]), unit='mi')

        similarity_matrix = np.zeros((n, n))

        for i in range(n):
            for j in range(n):
                if i == j:
                    similarity_matrix[i, j] = 0
                else:
                    similarity_matrix[i, j] = np.mean([demand[i], demand[j]]) * np.exp(-distance_matrix[i, j]/stdev**2)

        n = similarity_matrix.shape[0]
        similarity_matrix[np.tril_indices(n)] = 0

    return similarity_matrix

# Function to find and merge nodes with maximum similarity
def find_max_similarity_nodes(similarity_matrix):
    max_similarity = 0
    max_indices = (0, 0)

    # Iterate through the upper triangular part of the similarity matrix
    for i in range(len(similarity_matrix)):
        for j in range(i + 1, len(similarity_matrix)):
            if similarity_matrix[i, j] > max_similarity:
                max_similarity = similarity_matrix[i, j]
                max_indices = (i, j)

    return max_indices


### Initialization 

W = np.array(demand_df['Co-ordinates'].tolist())
demand = np.array(demand_df['Demand'].tolist())

W.shape

%%time
###----------------------AGGLOMERATIVE CLUSTERING TILL W = RQ----------------------###

while len(W) > Rq:
    ###----------------------GET MAX SIMILARITY INDICES ----------------------###
    similarity_matrix = get_similarity_matrix(W, demand, 100, 'euclidean')
    max_indices = find_max_similarity_nodes(similarity_matrix)
    print(max_indices)
    print(similarity_matrix[max_indices[0], max_indices[1]])

    ###----------------------COMPUTE W_new, demand_new AND THEN UPDATE W AND demand ACCORDINGLY----------------------###
    W_new = (demand[max_indices[0]]*W[max_indices[0]] + demand[max_indices[1]]*W[max_indices[1]]) / (demand[max_indices[0]] + demand[max_indices[1]])
    demand_new = demand[max_indices[0]] + demand[max_indices[1]]

    indices_to_remove = list(max_indices)
    mask = np.ones(len(W), dtype=bool)
    mask_demand = np.ones(len(demand), dtype=bool)
    mask[indices_to_remove] = False
    mask_demand[indices_to_remove] = False

    ###----------------------UPDATED COORDINATES (Wi MATRIX)----------------------###
    W = W[mask]
    W = list(W)
    W.append(W_new)
    W = np.array(W)

    ###----------------------UPDATED DEMAND (Di MATRIX)----------------------###
    demand = demand[mask]
    demand = np.append(demand, demand_new)
    
    

W.shape

### NOW THAT WE HAVE CLUBBED THE CUSTOMERS TOGETHER -> COMPUTE VQ~ (which is |VQ~| = Rq) SUCH THAT EVERY CUSTOMER HAS A MINIMUM DISTANCE VQ~

# q_df['Coordinates'] = q_df.apply(lambda x : [x['Latitude'], x['Longitude']], axis=1)

import numpy as np
from scipy.spatial import distance

vq_coordinates = q_df[['Latitude', 'Longitude']].values 
w_coordinates = W

# Compute distances
distances = distance.cdist(vq_coordinates, w_coordinates)

# Find the minimum distances for each set of points
min_distances_vq = np.min(distances, axis=1)
min_distances_w = np.min(distances, axis=0)

# Sort the indices based on the minimum distances
indices_vq = np.argsort(min_distances_vq)[:200]
indices_w = np.argsort(min_distances_w)[:200]

# Extract the points at minimum distances
vq_minimum_points = vq_coordinates[indices_vq]
w_minimum_points = w_coordinates[indices_w]


matching_indices_dict = {f'VQ_{vq_idx}': f'K_{w_idx}' for vq_idx, w_idx in zip(indices_vq, indices_w)}

### SERVING BETWEEN SECONDARY FACILITY TO CUSTOMERS 
matching_indices_dict

## Save the Vq to K dictionary to a pickle file
with open(result_path, 'wb') as f:
    pickle.dump(matching_indices_dict, f)

## Save the optimized customers and respective demands to .npy 
np.save(customer_path, W)
np.save(optimized_demand_path, demand)
